Refactor : 맡은 도메인 리팩토링
