+++ src/main/java/com/mypill/domain/address/dto/response/AddressResponse.java
@@ -1,11 +1,8 @@
 package com.mypill.domain.address.dto.response;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.mypill.domain.address.entity.Address;
-import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
 import lombok.Data;
-import lombok.Getter;
 
 @Data
 @Builder
