+++ src/main/java/com/mypill/domain/address/service/AddressService.java
@@ -6,7 +6,6 @@ import com.mypill.domain.address.repository.AddressRepository;
 import com.mypill.domain.member.entity.Member;
 import com.mypill.domain.member.service.MemberService;
 import com.mypill.global.AppConfig;
-import com.mypill.global.rq.Rq;
 import com.mypill.global.rsData.RsData;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
