+++ src/main/java/com/mypill/domain/comment/dto/CommentResponse.java
@@ -1,8 +1,7 @@
-package com.mypill.domain.post.dto;
+package com.mypill.domain.comment.dto;
 
 import com.mypill.domain.comment.entity.Comment;
 import com.mypill.domain.member.entity.Member;
-import com.mypill.domain.post.entity.Post;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
