+++ src/main/java/com/mypill/domain/comment/repository/CommentRepositoryImpl.java
@@ -1,17 +1,12 @@
 package com.mypill.domain.comment.repository;
 
+import com.mypill.domain.comment.dto.CommentResponse;
 import com.mypill.domain.comment.entity.QComment;
 import com.mypill.domain.member.entity.QMember;
-import com.mypill.domain.post.dto.CommentResponse;
-import com.mypill.domain.post.dto.PostResponse;
-import com.mypill.domain.post.entity.QPost;
 import com.querydsl.core.Tuple;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 import java.util.stream.Collectors;
