+++ src/main/java/com/mypill/domain/post/dto/PostRequest.java
@@ -1,13 +1,9 @@
 package com.mypill.domain.post.dto;
 
-import com.mypill.domain.post.entity.Post;
 import lombok.Data;
 
 @Data
 public class PostRequest {
-
     private String title;
     private String content;
-//    private MultipartFile image;
-
 }
