+++ src/main/java/com/mypill/global/springdoc/SpringDocConfig.java
@@ -11,8 +11,7 @@ import org.springframework.context.annotation.Configuration;
 @SecurityScheme(
         name = "bearerAuth",
         type = SecuritySchemeType.HTTP,
-        bearerFormat = "JWT",
-        scheme = "bearer"
+        scheme = "https"
 )
 public class SpringDocConfig {
 }
