+++ src/main/resources/templates/usr/post/detail.html
@@ -213,8 +213,6 @@
             commentContentElement.style.display = 'block';
             commentEditFormElement.style.display = 'none';
         }
-    </script>
-    <script>
         $(document).ready(function () {
             $('#newContent').on('keyup', function () {
                 $('#newContentCount').html("(" + $(this).val().length + " / 1000)");
